// src/app/features/animal-health/animal-treatment/models/open-treatment-response.model.ts
export interface OpenTreatmentCasesResponse {
  caseId: number;
  followUpNo: number;
  treatmentDate: string;
  treatmentRecordDate: string;
  vetName: string;
  samplingStatus: number;
  samplingStatusDesc: string;
  sampleDetails?: SampleDetailsCases[];
  radiologyDetails?: RadiologyDetailsCases[];
  diseaseDetails: DiseaseDetail[];
}

export interface DiseaseDetail {
  diseaseCd: number;
  diseaseDesc: string;
}

// Generated by https://quicktype.io

export interface SampleDetailsCases {
  sampleId: string;
  runSeqNo: number;
  courierId: null;
  diseaseCd: number;
  finalSampleResultValue: string;
  followUpNo: number;
  initialSampleResultValue: string;
  labCd: null;
  labCharges: null;
  modeOfTransport: null;
  onSpotTestCd: null;
  poolNoOfAnimals: null;
  receiptNo: null;
  sampleBarCd: null;
  sampleCollectionDate: null;
  sampleExaminationSubtypeCd: null;
  sampleExaminationTypeCd: null;
  sampleReport: null;
  sampleResult: number;
  sampleResultRecievedDate: null;
  sampleType: number;
  samplingStatus: number;
  sourceOriginCd: number;
  sourceOriginId: number;
  testImageUrl1: null;
  testImageUrl2: null;
  testRemarks: null;
  testingLocation: number;
}

// Generated by https://quicktype.io

export interface RadiologyDetailsCases {
  caseId: number;
  followUpNo: number;
  runSeqNo: number;
  radiologyReportDate: string;
  radiologyReportType: number;
  reportObservationsRemarks: null;
  testImageUrl1: null;
  testImageUrl2: null;
}